* Updated map server to partially 1203~1249
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@2292 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/mob.h
@@ -22,7 +22,7 @@ struct mob_skill {
 
 struct mob_db {
 	char name[NAME_LENGTH],jname[NAME_LENGTH];
-	int lv;
+	short lv;
 	int max_hp,max_sp;
 	int base_exp,job_exp;
 	int atk1,atk2;
